refactor: add services folder for api calls
